Refactored campaign options handling and spinner models.
Revised skill and financial option handling for clarity and functionality, including updates to skill target numbers, costs, and milestones. Introduced helper methods and refined constructors in `CampaignOptionsSpinner` for better tooltip and model management. Updated several fields from float to double to improve precision.

+++ MekHQ/src/mekhq/gui/dialog/campaignOptions/BiographyTab.java
@@ -1347,7 +1347,7 @@ public class BiographyTab {
         campaignOptions.setUseRandomClanPersonnelDeath(chkUseRandomClanPersonnelDeath.isSelected());
         campaignOptions.setUseRandomPrisonerDeath(chkUseRandomPrisonerDeath.isSelected());
         campaignOptions.setUseRandomDeathSuicideCause(chkUseRandomDeathSuicideCause.isSelected());
-        campaignOptions.setPercentageRandomDeathChance((int) spnPercentageRandomDeathChance.getValue());
+        campaignOptions.setPercentageRandomDeathChance((double) spnPercentageRandomDeathChance.getValue());
         for (final AgeGroup ageGroup : AgeGroup.values()) {
             campaignOptions.getEnabledRandomDeathAgeGroups().put(ageGroup,
                 chkEnabledRandomDeathAgeGroups.get(ageGroup).isSelected());
